Adjust hardware reports
Add software reports
