+++ src/main/java/com/systemscanner/api/controller/ReportController.java
@@ -13,7 +13,6 @@ import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 
-@CrossOrigin
 @RestController
 @AllArgsConstructor
 @RequestMapping("/reports")
