+++ src/main/java/com/systemscanner/api/controller/ScannerInstanceController.java
@@ -1,0 +1,33 @@
+package com.systemscanner.api.controller;
+
+import com.systemscanner.api.model.dto.NewScannerInstance;
+import com.systemscanner.api.model.projection.ScannerInstanceLight;
+import com.systemscanner.api.service.ScannerInstanceService;
+import lombok.AllArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.web.bind.annotation.*;
+
+import javax.validation.Valid;
+import java.util.Set;
+
+@RestController
+@AllArgsConstructor
+@RequestMapping("/scanner-instances")
+public class ScannerInstanceController {
+
+	private final ScannerInstanceService scannerInstanceService;
+
+	@GetMapping
+	public Set<ScannerInstanceLight> findAll(Authentication authentication) {
+		return this.scannerInstanceService.findAllForCurrentUser(authentication.getName());
+	}
+
+	@PostMapping
+	public ResponseEntity<ScannerInstanceLight> addInstance(@Valid @RequestBody NewScannerInstance newScannerInstance,
+															Authentication authentication) {
+		return this.scannerInstanceService.createNewInstance(newScannerInstance, authentication.getName())
+				.map(ResponseEntity::ok)
+				.orElse(ResponseEntity.notFound().build());
+	}
+}
