+++ src/main/java/com/systemscanner/api/controller/SingupController.java
@@ -4,11 +4,13 @@ import com.systemscanner.api.model.dto.SignupDTO;
 import com.systemscanner.api.model.projection.UserLight;
 import com.systemscanner.api.service.SignupService;
 import lombok.AllArgsConstructor;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import javax.validation.Valid;
 
-@CrossOrigin
 @RestController
 @AllArgsConstructor
 @RequestMapping("/auth")
@@ -20,4 +22,5 @@ public class SingupController {
 	public UserLight signup(@Valid @RequestBody SignupDTO signupDTO) {
 		return this.signupService.signup(signupDTO);
 	}
+
 }
