+++ src/main/java/com/systemscanner/api/model/mongo/Report.java
@@ -26,4 +26,5 @@ public class Report {
 
 	@JsonUnwrapped
 	private ReportInfo details;
+
 }
