+++ src/main/java/com/systemscanner/api/model/mongo/hardware/Win32ComputerSystem.java
@@ -1,4 +1,4 @@
-package com.systemscanner.api.model.mongo;
+package com.systemscanner.api.model.mongo.hardware;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
@@ -68,4 +68,5 @@ public class Win32ComputerSystem {
 
 	@JsonProperty("Workgroup")
 	private String workgroup;
+
 }
