+++ src/main/java/com/systemscanner/api/model/mongo/software/Software.java
@@ -1,0 +1,25 @@
+package com.systemscanner.api.model.mongo.software;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Software {
+
+	@JsonProperty("Name")
+	private String name;
+
+	@JsonProperty("Vendor")
+	private String vendor;
+
+	@JsonProperty("Version")
+	private String version;
+
+	@JsonProperty("InstallLocation")
+	private String installLocation;
+
+}
