+++ src/main/java/com/systemscanner/api/model/projection/ReportLight.java
@@ -9,4 +9,5 @@ public interface ReportLight {
 	String getScannerPid();
 
 	Instant getCreatedAt();
+
 }
