+++ src/main/java/com/systemscanner/api/model/projection/UserLight.java
@@ -5,4 +5,5 @@ public interface UserLight {
 	String getUsername();
 
 	String getEmail();
+
 }
