+++ src/main/java/com/systemscanner/api/service/ScannerInstanceService.java
@@ -7,7 +7,9 @@ import java.util.Optional;
 import java.util.Set;
 
 public interface ScannerInstanceService {
+
 	Set<ScannerInstanceLight> findAllForCurrentUser(String username);
 
 	Optional<ScannerInstanceLight> createNewInstance(NewScannerInstance instance, String username);
+
 }
