+++ src/main/java/com/systemscanner/api/service/SignupService.java
@@ -4,5 +4,7 @@ import com.systemscanner.api.model.dto.SignupDTO;
 import com.systemscanner.api.model.projection.UserLight;
 
 public interface SignupService {
+
 	UserLight signup(SignupDTO signupDTO);
+
 }
