+++ src/main/java/com/systemscanner/api/service/impl/ReportServiceImpl.java
@@ -37,4 +37,5 @@ public class ReportServiceImpl implements ReportService {
 		return this.mongoRepository.findById(id)
 				.filter(authorized);
 	}
+
 }
