+++ src/main/java/com/systemscanner/api/service/impl/ScannerInstanceServiceImpl.java
@@ -52,4 +52,5 @@ public class ScannerInstanceServiceImpl implements ScannerInstanceService {
 				.securityKey(instance.getToken())
 				.build();
 	}
+
 }
