+++ src/main/java/com/systemscanner/api/service/impl/ScannerServiceImpl.java
@@ -5,20 +5,34 @@ import com.systemscanner.api.model.mongo.Report;
 import com.systemscanner.api.model.mongo.ReportInfo;
 import com.systemscanner.api.repository.jpa.ScannerInstanceRepository;
 import com.systemscanner.api.repository.mongo.ReportMongoRepository;
-import com.systemscanner.api.service.RatService;
-import lombok.AllArgsConstructor;
+import com.systemscanner.api.service.ScannerService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.time.Instant;
 import java.util.Optional;
+import java.util.UUID;
 
 @Service
-@AllArgsConstructor
-public class RatServiceImpl implements RatService {
+@RequiredArgsConstructor
+public class ScannerServiceImpl implements ScannerService {
 
 	private final ScannerAuthenticator scannerAuthenticator;
 	private final ReportMongoRepository reportMongoRepository;
 	private final ScannerInstanceRepository scannerInstanceRepository;
+
+	@Value("${scanner.secret}")
+	private String secret;
+
+	@Override
+	public Optional<String> registerInstance(String scannerSecret) {
+		return Optional.of(scannerSecret)
+				.filter(secret::equals)
+				.map(authorized -> this.createInstance())
+				.map(this.scannerInstanceRepository::save)
+				.flatMap(this.scannerAuthenticator::authenticate);
+	}
 
 	@Override
 	public Optional<Report> createReport(ReportInfo reportInfo, String scannerToken) {
@@ -28,6 +42,13 @@ public class RatServiceImpl implements RatService {
 				.map(reportMongoRepository::save);
 	}
 
+	private ScannerInstance createInstance() {
+		return ScannerInstance.builder()
+				.pid(UUID.randomUUID().toString())
+				.securityKey(UUID.randomUUID().toString())
+				.build();
+	}
+
 	private Report buildReport(ReportInfo reportInfo, ScannerInstance scanner) {
 		return Report.builder()
 				.scannerPid(scanner.getPid())
@@ -35,4 +56,5 @@ public class RatServiceImpl implements RatService {
 				.details(reportInfo)
 				.build();
 	}
+
 }
