+++ src/main/java/com/systemscanner/api/utils/StringUtils.java
@@ -3,7 +3,9 @@ package com.systemscanner.api.utils;
 public interface StringUtils {
 
 	String BLANK = "";
+
 	String COMMA = ",";
+
 	String COLON = ":";
 
 }
