+++ src/main/resources/application.properties
@@ -15,7 +15,11 @@ spring.jpa.dialect = org.hibernate.dialect.PostgreSQL95Dialect
 
 spring.jpa.hibernate.ddl-auto = validate
 
+spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation = true
+
 server.servlet.context-path = /api/v1
 
 api.token.max-age = 120
 api.token.secret = some-not-secure-secret
+
+scanner.secret = ${SCANNER_SECRET}
